more improvements
Today in `PreSweepCallback` we would invalidate the guard if the underline type is not marked. This would always fail the `CheckIfTypeIsEquivalent`
because we first check if guard value is valid before proceeding further. Thus we would `BailOutFailedEquivalentTypeCheck`. However we could
have invalid guard value but if there are alive types present in equivTypes[], we could use them to check if `type` in question matches any of them
and if yes, update the guard value, so next time JIT code will be able to take advantage of it next time.

On instrumenting code, noticed that there are high frequency that this happens (at least in acmeair) where we have valid equivTypes[], but we don'take
take its advantage just because guard value is invalid. This gives approx. 2.5 % win in acmeair. Other benchmarks are flat.

With that I reverted my previous optimization of zeroing out equivTypes[] is guard is invalidated.