Fixed bug related to cross context
Earlier inside `CheckIfTypeIsEquivalent` we verified if scriptContext of type we are searching and
that of guard type matches. If not, we would bailout. However in last commit I added additional
condition to do this check only if guard has not been invalidated during sweep. So we have a scenario
where guard has been invalidated while sweeping and a new type from different scriptContext comes in
`CheckIfTypeIsEquivalent`. We skip checking for type's scriptContext and then reset the guard with this
type (from slow path). This would fail some of the assumptions made further in JIT code with regards
to scriptContext of type. Fixed it by checking the scriptContext of type with that of the cached types.
If they don't match return false which will bailout.

Also noticed we were unnecessarily looping in cache eviction policy code path to search for null cached types.
Since we already do that, we can just directly evict the right candidate.

Test : In progress