Fixed bug to stop using wrong types.
In my earlier implementation, I was distinguishing types based on slotCapacity. However other types (e.g. `TypeIds_Object` and `TypeIds_Uint8Array`) might are `DynamicType`, but sharing their type will be wrong. To address that I am categorizing slots based on type and typeId. Slot `TypeOfPrototypeObjectInlined` is used if object's type is `DynamicType` and it has 'TypeIds_Object` as its typeid. This would work out of box in CreateObject scenario because it is only used for DynamicType having typeid as `TypeIds_Object`. Slot `TypeOfPrototypeObjectDictionary` would be used to cache everything else. As name says, value of this slot would be dictionary where key is combination of type and the typeId. For instance, DyanmicType+TypeIds_Date, DynamicType+TypeIds_Uint8Array and JsrtExternalType+TypeIds_Object will all have seperate entries in dictionary. Value of the dictionary would be actual type we want to cache.

With this, we will not use type sharing for ExternalTypes.