CheckIfTypeIsEquivalent improvement
If we get a non-shared type and if cache is full, we will consider that there is no point in checking for equivalence and we would just go to slow path / helper for doing the operation. However if shared type comes and cache is full, we will first try to evict non-shared type before evicting shared types. Empty slots takes first priority over non-shared types in eviction.