Arguments symbol is uninitialized when a function definition with the name arguments occur in the body in non-split scope
When a function definition with the name arguments occurs in the body it
makrs the function as arguments creation is not needed. The arguments is
initialized only at the beginning of the body. So when arguments is used
in the param scope it will be unitialized. Also if arguments symbol is
captured in the param scope we should split the scope as it can be
overwritten in the body.