Do not duplicate the formals in the body if there is no var declaration shadowing it
Currently we create a duplicate symbol for every formal in the body. As
per the spec this is not needed. New symbol is needed only when there is a
var declaration in the body that shadows the formal. If no vars shadow a
formal then it is always accessed from the param scope itself. If a var
shadows a formal then we follow the existing path to copy over the value
from the formal to the corresponding symbol in the body.