ForInObjectEnumerator optimizations
Profiled InitForInEnumerator opcode and fast path
Stack Allocate ForInObjectEnumerator
Reduce forin size
Remove JavascriptEnumeratorIterator, and change JavascriptDispatch to cache the current index so ForInObjectEnumerator doesn't have to.

If prototype doesn't have any enumerable property, we don't have to keep track of shadowing.  So just ask the object enumerator to return enumerable property only, and the for in object enumerator and the jit fast path doesn't have to check.

Instead of checking multiple condition, tnitialize a single bool to indicate whether we can use the fast path, that combines the check for:
- Whether we have a null object property enumerator
- Whether we have a prefix/array enumerator
- Whether we need to keep track of shadowing because prototype has enumerable properties
- Whether we are using cached data